Update isUnexpected.ts file
