+++ packages/cadl-rlc-test/cadl-output/src/clientDefinitions.ts
@@ -2,6 +2,8 @@ import { Client, StreamableMethod } from "@azure-rest/core-client";
 
 /** Contains operations for Pets operations */
 export interface PetsOperations {
+    /** Delete a pet. */
+    delete(petId: number, options: Options): StreamableMethod<>;
     /** Returns a pet. Supports eTags. */
     read(petId: number, options: Options): StreamableMethod<>;
     create(options: Options): StreamableMethod<>;
@@ -12,7 +14,9 @@ export interface ListPetToysResponseOperations {
     list(petId: string, options: Options): StreamableMethod<>;
 }
 
-export interface PetsRead {
+export interface PetsDelete {
+    /** Delete a pet. */
+    delete(options: Options): StreamableMethod<>;
     /** Returns a pet. Supports eTags. */
     get(options: Options): StreamableMethod<>;
 }
@@ -26,8 +30,8 @@ export interface ListPetToysResponseList {
 }
 
 export interface Routes {
-    /** Resource for '/pets/\{petId\}' has methods for the following verbs: get */
-    (path: "/pets/{petId}", petId: number): PetsRead;
+    /** Resource for '/pets/\{petId\}' has methods for the following verbs: delete, get */
+    (path: "/pets/{petId}", petId: number): PetsDelete;
     /** Resource for '/pets' has methods for the following verbs: post */
     (path: "/pets"): PetsCreate;
     /** Resource for '/pets/\{petId\}/toys' has methods for the following verbs: get */
