+++ packages/cadl-rlc-test/cadl-output/src/isUnexpected.ts
@@ -1,0 +1,3 @@
+import "./responses";
+
+const responseMap: Record<string, string[]> = {"DELETE /pets/{petId}":["200"],"GET /pets/{petId}":["200"],"POST /pets":["200"],"GET /pets/{petId}/toys":["200"]};
