+++ packages/rlc-codegen/src/index.ts
@@ -5,6 +5,7 @@ export * from "./interfaces.js";
 export * from "./generateClientDefinitions.js";
 export * from "./generateSchemaType.js";
 export * from "./helpers/shortcutMethods.js";
-export * from "./helpers/nameUtils.js"
+export * from "./helpers/nameUtils.js";
 export * from "./buildResponseTypes.js";
 export * from "./buildParameterTypes.js";
+export * from "./buildIsUnexpectedHelper.js";
